Strip VMEM permission before handing caps to free
When mrs was first written, it assumed that all underlying allocators
would use only the address of the capability passed to free, rederiving
a new capability to the underlying allocation either at free() or as
part of the subsequent malloc() that re-issued the address space in
question.

While a nicely defensive posture on the part of the allocator, it is
probably not the behavior we want to mandate.  Towards that end, strip
the VMEM permission we're using to preserve our quarantine enumeration
across revocation before we hand the pointer back to malloc, lest a
VMEM-bearing capabilty escape to the allocator's client.